Set layout on all buttons and panel
Layout was set using flexbox. Every button is on a div, in order to
adjust its size. Some buttons are bigger resembling a common calculator
layout pattern. All logic on the buttons is yet to be made with JavaScript.
Also, the labels for IDs and classes on HTML needs improvement for better
clarity.