Add logic to add values and operate
Hitting any opperator when an opperator has already been assigned
will trigger the current evaluation. Enter button also works.
Floating point numbers are rounded to up to 2 digits in order.
to not overflow the screen

Code needs further readability improvements, with new functions
and new function names, especially on the buttons event listener
Zeros at left should also be replaced with other digits, but
they are stacking.

A suggestion for a new feature on future versions is a backspace
button to undo the button hit (digit or operator)