Fix zeroes stacking at left on first value
A lone zero will always be replaced with the next digit, following the standard
notation for the decimal system. This fix happens on the buttons event listener.
Althought, the second value still stacks zeroes on the left.

There are a few bugs still occurring such as: enter button without first or second
number values; first value inputed by the user is an opperator; pressing two
operators in a row; pressing two points in a row; pressing two clear in a row;