Fix zeroes stacking on second value
Zeroes at left of second value no longer stack. Instead,
they are replaced with the next number input. Additionally,
minor bugs on clear button interaction were corrected.

Trying to evaluate an expression with a missing value,
on pressing two consecutive operators, for instance, will
assume that one of the values is zero, and that is unintended.